Fix part related with history_points

+++ src/input.jl
@@ -49,9 +49,9 @@ Base.@kwdef struct TOMLInput{Model}
 end
 
 Base.@kwdef mutable struct TOMLOutput
-    directory      :: String      = ""
-    num_data       :: Int         = 100
-    history_points :: Vector{Int} = Int[]
+    directory      :: String          = ""
+    num_data       :: Int             = 100
+    history_points :: Vector{Float64} = Float64[]
 end
 
 Base.@kwdef struct TOMLNewmarkBeta
